assignment2/scoring/requirements.txt
streamlit
joblib
lightgbm
pandas
Pillow